move make_move function into Move module
