+++ tests/test_move_validation.py
@@ -1,8 +1,7 @@
 import pytest
 
 from utahchess.board import Board
-from utahchess.legal_moves import make_move
-from utahchess.move import REGULAR_MOVE, Move
+from utahchess.move import REGULAR_MOVE, Move, make_move
 from utahchess.move_candidates import get_king_move_candidates, get_pawn_move_candidates
 from utahchess.move_validation import is_checkmate, is_valid_move
